requirement.txt
langchain
openai
pathspec
python-dotenv
streamlit
streamlit_chat
tiktoken
qdrant-client
langchain-community